add license compliance metrics

+++ monitoring/node_state.go
@@ -17,10 +17,10 @@ type NodeState struct {
 // NewNodeState returns the status for each node in the Vertica cluster.
 func NewNodeState(db *sqlx.DB) []NodeState {
 	sql := `
-	SELECT 
-		node_id, 
-		node_name, 
-		(node_state='UP')::INT node_state 
+	SELECT
+		node_id,
+		node_name,
+		(node_state='UP')::INT node_state
 	FROM v_catalog.nodes`
 
 	nodeState := []NodeState{}
@@ -33,12 +33,12 @@ func NewNodeState(db *sqlx.DB) []NodeState {
 }
 
 // ToMetric converts NodeState to a Map.
-func (ns NodeState) ToMetric() map[string]int {
-	metrics := map[string]int{}
+func (ns NodeState) ToMetric() map[string]float64 {
+	metrics := map[string]float64{}
 
 	id := fmt.Sprintf("node_id=%q", ns.NodeID)
 	name := fmt.Sprintf("node_name=%q", ns.NodeName)
-	metrics[fmt.Sprintf("vertica_node_state{%s, %s}", id, name)] = ns.NodeState
+	metrics[fmt.Sprintf("vertica_node_state{%s, %s}", id, name)] = float64(ns.NodeState)
 
 	return metrics
 }
