+++ monitoring/pool_usage.go
@@ -20,11 +20,11 @@ type PoolUsage struct {
 func NewPoolUsage(db *sqlx.DB) []PoolUsage {
 	sql := `
 	SELECT
-		node_name, 
-		pool_name, 
+		node_name,
+		pool_name,
 		memory_inuse_kb,
-		general_memory_borrowed_kb, 
-		running_query_count 
+		general_memory_borrowed_kb,
+		running_query_count
 	FROM resource_pool_status;
 	`
 
@@ -38,14 +38,14 @@ func NewPoolUsage(db *sqlx.DB) []PoolUsage {
 }
 
 // ToMetric converts PoolUsage to a Map.
-func (usage PoolUsage) ToMetric() map[string]int {
-	metrics := map[string]int{}
+func (usage PoolUsage) ToMetric() map[string]float64 {
+	metrics := map[string]float64{}
 
 	node := fmt.Sprintf("node_name=%q", usage.NodeName)
 	pool := fmt.Sprintf("pool_name=%q", usage.PoolName)
-	metrics[fmt.Sprintf("vertica_pool_memory_inuse_kb{%s, %s}", node, pool)] = usage.MemoryInUseKB
-	metrics[fmt.Sprintf("vertica_pool_memory_borrowed_kb{%s, %s}", node, pool)] = usage.GeneralMemoryBorrowedKB
-	metrics[fmt.Sprintf("vertica_pool_running_query_count{%s, %s}", node, pool)] = usage.RunningQueryCount
+	metrics[fmt.Sprintf("vertica_pool_memory_inuse_kb{%s, %s}", node, pool)] = float64(usage.MemoryInUseKB)
+	metrics[fmt.Sprintf("vertica_pool_memory_borrowed_kb{%s, %s}", node, pool)] = float64(usage.GeneralMemoryBorrowedKB)
+	metrics[fmt.Sprintf("vertica_pool_running_query_count{%s, %s}", node, pool)] = float64(usage.RunningQueryCount)
 
 	return metrics
 }
